Parser bug fix. Default type for Strong is unknown
In the `strong-types` repo file `src/map/parser.ts` on line 136
the one checking if `child.typeId === 'StrongType'`,
there should be null check.

The optional chain operator `child?.typeId` will make sure
`child` is not `undefined` or `null`.

We shouldn't do a null check before this point
because `null` is a valid value for `parseKey` to handle.